pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ask-finance"
version = "0.1.0"
description = "Ask Finance: finance Q&A agent prototype (Gradio + Gemini on Vertex)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "gradio>=4.44.0",
  "google-genai>=1.0.0",
  "pandas>=2.0.0",
  "pydantic>=2.0.0",
  "matplotlib>=3.7.0",
  "openpyxl>=3.1.0",
  "python-pptx>=0.6.21",
]

[project.optional-dependencies]
dev = ["ruff>=0.4.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
